u-boot-marvell/tools/env/fw_env.c - make an implicit #define override explicit
This manifests because the u-boot-marvell/tools/Makefile now includes
the env subirectory.  It isn't directly something to do with the
ionicsplug patches, it happens anyway when env is included.

Therefore, and since the #define of CONFIG_ENV_SIZE is within a .c
file, I am concluding that it is just a coincidence it is named the
same as the one in sheevaplug.h and it is in fact independent.

Therefore I'm undefining CONFIG_ENV_SIZE before redefining it, to make
this explicit and squelch the warning.